+++ src/cli/show.rs
@@ -1,9 +1,13 @@
-use crate::{cli::Subcommand, util::Directory, GlobalArgs};
+use crate::{
+    cli::Subcommand, collection::CollectionFile, config::Config, db::Database,
+    util::paths::DataDirectory, GlobalArgs,
+};
 use async_trait::async_trait;
 use clap::Parser;
-use std::process::ExitCode;
+use serde::Serialize;
+use std::{borrow::Cow, path::Path, process::ExitCode};
 
-/// Show meta information about slumber
+/// Print meta information about Slumber (config, collections, etc.)
 #[derive(Clone, Debug, Parser)]
 pub struct ShowCommand {
     #[command(subcommand)]
@@ -12,16 +16,50 @@ pub struct ShowCommand {
 
 #[derive(Copy, Clone, Debug, clap::Subcommand)]
 enum ShowTarget {
-    /// Show the directory where slumber stores data and log files
-    Dir,
+    /// Print the path of all directories/files that Slumber uses
+    Paths,
+    /// Print loaded configuration
+    Config,
+    /// Print current request collection
+    Collection,
 }
 
 #[async_trait]
 impl Subcommand for ShowCommand {
-    async fn execute(self, _global: GlobalArgs) -> anyhow::Result<ExitCode> {
+    async fn execute(self, global: GlobalArgs) -> anyhow::Result<ExitCode> {
         match self.target {
-            ShowTarget::Dir => println!("{}", Directory::root()),
+            ShowTarget::Paths => {
+                let collection_path =
+                    CollectionFile::try_path(global.collection);
+                println!("Data directory: {}", DataDirectory::root());
+                println!("Log file: {}", DataDirectory::log());
+                println!("Config: {}", Config::path());
+                println!("Database: {}", Database::path());
+                println!(
+                    "Collection: {}",
+                    collection_path
+                        .as_deref()
+                        .map(Path::to_string_lossy)
+                        .unwrap_or_else(|error| Cow::Owned(error.to_string()))
+                )
+            }
+            ShowTarget::Config => {
+                let config = Config::load()?;
+                println!("{}", to_yaml(&config));
+            }
+            ShowTarget::Collection => {
+                let collection_path =
+                    CollectionFile::try_path(global.collection)?;
+                let collection_file =
+                    CollectionFile::load(collection_path).await?;
+                println!("{}", to_yaml(&collection_file.collection));
+            }
         }
         Ok(ExitCode::SUCCESS)
     }
+}
+
+fn to_yaml<T: Serialize>(value: &T) -> String {
+    // Panic is intentional, indicates a wonky bug
+    serde_yaml::to_string(value).expect("Error serializing")
 }