+++ src/config.rs
@@ -1,28 +1,24 @@
 use crate::{
     tui::input::{Action, InputBinding},
-    util::{parse_yaml, Directory, ResultExt},
+    util::{
+        parse_yaml,
+        paths::{DataDirectory, FileGuard},
+        ResultExt,
+    },
 };
 use anyhow::Context;
 use indexmap::IndexMap;
-use serde::Deserialize;
-use std::{
-    fs,
-    path::{Path, PathBuf},
-};
+use serde::{Deserialize, Serialize};
+use std::fs;
 use tracing::info;
 
 /// App-level configuration, which is global across all sessions and
 /// collections. This is *not* meant to modifiable during a session. If changes
 /// are made to the config file while a session is running, they won't be
 /// picked up until the app restarts.
-#[derive(Debug, Deserialize)]
+#[derive(Debug, Serialize, Deserialize)]
 #[serde(default, deny_unknown_fields)]
 pub struct Config {
-    /// The path that the config was loaded from, or tried to be loaded from if
-    /// the file didn't exist
-    #[serde(skip)]
-    path: PathBuf,
-
     /// TLS cert errors on these hostnames are ignored. Be careful!
     #[serde(default)]
     pub ignore_certificate_hosts: Vec<String>,
@@ -42,10 +38,10 @@ impl Config {
     /// deserialization failed. This is *not* async because it's only run during
     /// startup, when all operations are synchronous.
     pub fn load() -> anyhow::Result<Self> {
-        let path = Directory::root().create()?.join(Self::FILE);
+        let path = Self::path().create_parent()?;
         info!(?path, "Loading configuration file");
 
-        let mut config = match fs::read(&path) {
+        match fs::read(&path) {
             Ok(bytes) => parse_yaml::<Self>(&bytes)
                 .context(format!("Error loading configuration from {path:?}"))
                 .traced(),
@@ -59,22 +55,18 @@ impl Config {
                 );
                 Ok(Self::default())
             }
-        }?;
-
-        config.path = path;
-        Ok(config)
+        }
     }
 
-    /// The path where configuration is stored
-    pub fn path(&self) -> &Path {
-        &self.path
+    /// Path to the configuration file
+    pub fn path() -> FileGuard {
+        DataDirectory::root().file(Self::FILE)
     }
 }
 
 impl Default for Config {
     fn default() -> Self {
         Self {
-            path: PathBuf::default(),
             ignore_certificate_hosts: Vec::new(),
             preview_templates: true,
             input_bindings: IndexMap::default(),